+++ src/Discovery.js
@@ -93,6 +93,30 @@ module.exports = {
     return addresses
   },
 
+  getMultipleIps: async () => {
+    
+    const addresses = []
+    let name
+    let ifaces
+    let iface
+    const interfaces = networkInterfaces()
+    for (name in interfaces) {
+      // eslint-disable-next-line no-prototype-builtins
+      if(interfaces.hasOwnProperty(name)) {
+        ifaces = interfaces[name]
+        if(!/(loopback|internal)/gi.test(name)) {
+          for (let i = 0; i < ifaces.length; i++) {
+            iface = ifaces[i]
+            if (iface.family === 'IPv4' &&  !iface.internal && iface.address !== '127.0.0.1') {
+              addresses.push(iface.address)
+            }
+          }
+        }
+      }
+    }
+    return addresses
+  },
+
   getIpStephan: async () => {
     const ifaces = networkInterfaces()